# swagger.json
{"openapi":"3.0.0","paths":{"/wallets":{"post":{"operationId":"WalletController_createWallets","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWalletDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWalletsResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpMessageDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpMessageDto"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpMessageDto"}}}}},"tags":["Wallets"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpMessageDto"}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}},"additionalProperties":true}}}}}}}},"tags":["Health"]}}},"info":{"title":"LLC Crypto API","description":"The API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"HttpMessageDto":{"type":"object","properties":{"message":{"oneOf":[{"type":"array","items":{"type":"string"}},{"type":"string"}]},"success":{"type":"boolean","example":true}},"required":["message","success"]},"CreateWalletDto":{"type":"object","properties":{"wallets":{"description":"The wallets to create","example":[{"currency":"BTC","address":"1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa","privateKey":"1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa"}],"type":"array","items":{"type":"array"}}},"required":["wallets"]},"CreateWalletsResponseDto":{"type":"object","properties":{"message":{"oneOf":[{"type":"array","items":{"type":"string"}},{"type":"string"}]},"success":{"type":"boolean","example":true},"data":{"type":"object","description":"The data of the response","example":{"invalidWallets":[]},"nullable":true}},"required":["message","success"]}}}}